functions/graph_gen/requirements.txt
networkx
py-cpuinfo